# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SSH Commander API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alert-channels:
    get:
      tags: [alerts]
      summary: List notification channels (secrets masked)
      operationId: listAlertChannels
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationChannel"
    post:
      tags: [alerts]
      summary: Create a notification channel (admin)
      operationId: createAlertChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
  /alert-channels/{id}:
    put:
      tags: [alerts]
      summary: Update a notification channel (admin; omitted secrets are kept)
      operationId: updateAlertChannel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertChannelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [alerts]
      summary: Delete a notification channel (admin)
      operationId: deleteAlertChannel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /alert-channels/{id}/test:
    post:
      tags: [alerts]
      summary: Send a test message through a channel (admin)
      operationId: testAlertChannel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Delivery attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertTestResult"
  /alert-rules:
    get:
      tags: [alerts]
      summary: List alert rules
      operationId: listAlertRules
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRule"
    post:
      tags: [alerts]
      summary: Create an alert rule (admin)
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
  /alert-rules/{id}:
    put:
      tags: [alerts]
      summary: Update an alert rule (admin)
      operationId: updateAlertRule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [alerts]
      summary: Delete an alert rule (admin)
      operationId: deleteAlertRule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /alert-events:
    get:
      tags: [alerts]
      summary: Alert delivery history (newest first)
      operationId: listAlertEvents
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer }
        - name: offset
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Page of alert events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertEventPage"

  # ─── Audit Log (2.0.0) ───────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    NotificationChannel:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [telegram, email, webhook] }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
          description: Sanitized per-type config; secrets are replaced by *Set booleans
        lastUsedAt: { type: ["string", "null"], format: date-time }
        lastError: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, type, enabled, config, createdAt]
    AlertChannelInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, enum: [telegram, email, webhook] }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
          description: "telegram: {botToken, chatId}; email: {host, port, secure, username, password, from, to}; webhook: {url, secret}"
      required: [name, type, config]
    AlertChannelUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
          description: Omitted or empty secret fields keep their stored values
    AlertRule:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        eventTypes:
          type: array
          items:
            type: string
            enum: [device_down, device_up, job_failed, job_completed, schedule_failed, backup_failed, drift_detected, upgrade_completed, upgrade_failed]
        routerIds:
          type: array
          items: { type: integer }
        groupIds:
          type: array
          items: { type: integer }
        channelIds:
          type: array
          items: { type: integer }
        cooldownMinutes: { type: integer }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, eventTypes, routerIds, groupIds, channelIds, cooldownMinutes, enabled, createdAt]
    AlertRuleInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        eventTypes:
          type: array
          items:
            type: string
            enum: [device_down, device_up, job_failed, job_completed, schedule_failed, backup_failed, drift_detected, upgrade_completed, upgrade_failed]
        routerIds:
          type: array
          items: { type: integer }
        groupIds:
          type: array
          items: { type: integer }
        channelIds:
          type: array
          items: { type: integer }
        cooldownMinutes: { type: integer }
        enabled: { type: boolean }
      required: [name, eventTypes, channelIds]
    AlertRuleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        eventTypes:
          type: array
          items:
            type: string
            enum: [device_down, device_up, job_failed, job_completed, schedule_failed, backup_failed, drift_detected, upgrade_completed, upgrade_failed]
        routerIds:
          type: array
          items: { type: integer }
        groupIds:
          type: array
          items: { type: integer }
        channelIds:
          type: array
          items: { type: integer }
        cooldownMinutes: { type: integer }
        enabled: { type: boolean }
    AlertEvent:
      type: object
      properties:
        id: { type: integer }
        ruleId: { type: ["integer", "null"] }
        ruleName: { type: string }
        eventType: { type: string }
        subject: { type: string }
        message: { type: string }
        status: { type: string, enum: [sent, failed, suppressed] }
        error: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, ruleName, eventType, subject, message, status, createdAt]
    AlertEventPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AlertEvent"
        total: { type: integer }
      required: [items, total]
    AlertTestResult:
      type: object
      properties:
        success: { type: boolean }
        error: { type: ["string", "null"] }
      required: [success]

    # ─── Audit Log (2.0.0) ──────────────────────────────────────────────
...